Add style to .content & .button-container & button
add margin and padding to .content.
add margin and padding to .button-container.

to button, add display: block to move it to a new line.
but when i do it, .button container is not centered anymore.
so i add text-align: center to .button-container.
need to review this steps as there should be a better solution.

also add margin and padding to button.